Abbreviated SLES displayName
For SLES images in the preview portal, a too long display name was used which
made it hard to distinguish the different SLES flavors. Formerly, when
looking at the overview, every SLES image was labeled `SUSE Linux Enterprise Serve...`.
This is fixed now by using the abbreviation 'SLES' instead of the full name.