Delete CentOS_7.0.json
We’ve identified a platform bug that impacts CentOS 7.0/7.1 kernels.  This can cause the VM to hang when the Host OS is updated until the customer reboots the VM.  It won’t happen 100% of the time, but customers with larger deployments of these old kernels will see some subset of their VMs hang when updates roll out, which is a bad experience.  Unfortunately, due to the nature of CentOS, even if RHEL backports the patch to those older kernels it won’t help us because CentOS essentially “freezes” a release as soon as the next minor version is released.